refactor(my-diary): change components names
affects: @practice/my-diary

ISSUES CLOSED: #59
