+++ practice/my-diary/src/components/post-list-item/PostListItem.tsx
