+++ practice/my-diary/src/components/post-status-filter/index.tsx
@@ -1,3 +1,3 @@
-import PostStatusFilter from './Post-status-filter'
+import PostStatusFilter from './PostStatusFilter'
 
 export default PostStatusFilter
